Misc Changes to FNV
-Added '\n' new line to Test_Framework FNV0 Test for formatting
purposes.

-Set Int64 to Int32 is was meant for FNV-0 algorithm function.

-Update to Solution's FNV-Hash.dll reference.

+++ Source/Net-Hash/FNV-1-Hash/FNV.cs
@@ -11,7 +11,7 @@ namespace FNV_Hash
 
         public static string fnv032Hash(byte[] message)
         {
-            Int64 hash = 0;
+            Int32 hash = 0;
 
             foreach(byte data in message)
             {
